requirements.txt
argh==0.26.2
click==6.7
Flask==0.12.2
Flask-API==1.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
nose-watch==0.9.2
pathtools==0.1.2
py-dateutil==2.2
PyYAML==3.12
six==1.11.0
watchdog==0.8.3
Werkzeug==0.13
